backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
openpyxl
pydantic
supabase
asyncpg
python-dotenv
python-jose
httpx 